+++ NEWS.md
@@ -1,3 +1,9 @@
+# ggdemetra 0.2.7
+
+- deprecated function `y_forecast()` removed.
+
+- `calendar()`, `calendaradj()` and `autoplot()` correction with TRAMO-SEATS models.
+
 # ggdemetra 0.2.6
 
 * `siratio()` correction with TRAMO-SEATS `"jSA"` models.